+++ lib/providers/query/restquery.js
@@ -9,7 +9,7 @@ class RestQuery extends Query {
     if(!this.oql) {
       throw new Error('Cannot query without oql!');
     }
-    let appx = this.parentProvider.config;
+    let appx = this.bridge.getConfig();
     if(!appx) {
       throw new Error('Illegal state could not resolve queryProvider REST config');
     }
