+++ test/providers/fetch-request.spec.js
@@ -1,9 +1,9 @@
 /*jshint expr: true*/
 const expect = require('chai').expect;
-const Bridge = require('../lib/providers/bridge');
-const BaseFetchRequest = require('../lib/providers/fetch/fetch-request');
-const LocalFetchRequest = require('../lib/providers/fetch/local-fetch-request');
-const RestFetchRequest = require('../lib/providers/fetch/rest-fetch-request');
+const Bridge = require('../../lib/providers/bridge');
+const BaseFetchRequest = require('../../lib/providers/fetch/fetch-request');
+const LocalFetchRequest = require('../../lib/providers/fetch/local-fetch-request');
+const RestFetchRequest = require('../../lib/providers/fetch/rest-fetch-request');
 const sync = require('synchronize');
 
 describe('Base FetchRequest', () => {
