+++ test/providers/script-engine-api-request.spec.js
@@ -1,8 +1,8 @@
 /*jshint expr: true*/
 const expect = require('chai').expect;
 const sync = require('synchronize');
-const RestScriptEngineApiRequest = require('../lib/providers/api/rest-script-engine-api-request');
-const conf = require('../lib/config-loader').readConfig();
+const RestScriptEngineApiRequest = require('../../lib/providers/api/rest-script-engine-api-request');
+const conf = require('../../lib/config-loader').readConfig();
 
 
 describe.skip('RestScriptEngineApiRequest', () => {
